打者指標を計算用クラスと保存用クラスに分けた
